# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSphere Go API
  version: 1.0.0
  description: Contratos principales expuestos por el API Gateway.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/auth/register:
    post:
      summary: Registrar usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201': { description: Usuario registrado y JWT generado }
        '409': { description: Usuario o correo duplicado }
  /api/auth/login:
    post:
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emailOrUsername, password]
              properties:
                emailOrUsername: { type: string }
                password: { type: string }
      responses:
        '200': { description: JWT y datos del usuario }
        '401': { description: Credenciales inválidas }
  /api/users/me:
    get:
      summary: Obtener perfil autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Usuario y canal }
  /api/channels:
    post:
      summary: Crear canal y convertir al usuario en creador
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201': { description: Canal creado y JWT actualizado }
...